moved files into views
