+++ src/builderMockups/Main.java
@@ -4,6 +4,8 @@ import java.awt.EventQueue;
 
 import javax.swing.UIManager;
 
+import view.BuilderView;
+
 public class Main {
 
 	public static void main(String[] args) {
