+++ src/controllers/CreateLevelBtnController.java
@@ -3,6 +3,7 @@ package controllers;
 import java.awt.event.ActionEvent;
 
 import view.LevelTypeSelectView;
+import view.BuilderView;
 
 public class CreateLevelBtnController implements java.awt.event.ActionListener {
 	int nextLevel;
