+++ src/view/AbstractPieceGroupView.java
@@ -1,4 +1,4 @@
-package builderMockups;
+package view;
 
 import java.awt.Dimension;
