+++ src/view/SelectedPieceView.java
@@ -1,4 +1,4 @@
-package builderMockups;
+package view;
 
 import java.awt.Dimension;
 import javax.swing.JPanel;
